tpm2: export functions to make a credential directly with the library
Export TPM2_Startup(), TPM2_LoadExternal_SW(), TPM2_MakeCredential_SW() and
TPM2_FlushContext(), so that a caller can initialize the TPM, load the
public part of the EK, make a credential and flush the loaded key.

The _SW suffix means that the caller has to pass marshalled parameters
(pointers cannot be passed directly because libtpms and libibmtss define
structures with different array size).

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>